+++ lib/tappay/jko_pay/pay.rb
@@ -4,7 +4,7 @@ module Tappay
   module JkoPay
     class Pay < PaymentBase
       def endpoint_url
-        Endpoints::Payment.pay_by_prime_url
+        Tappay::Endpoints::Payment.pay_by_prime_url
       end
 
       private
@@ -21,16 +21,43 @@ module Tappay
         [:prime, :frontend_redirect_url, :backend_notify_url, :cardholder]
       end
 
+      def validate_options!
+        super
+        validate_result_url_format!
+      end
+
+      def validate_result_url_format!
+        # First validate that if result_url is provided, it's a hash with required fields
+        if options.key?(:result_url)
+          raise ValidationError, "result_url must be a hash" unless options[:result_url].is_a?(Hash)
+
+          result_url = options[:result_url]
+          required_fields = %w[frontend_redirect_url backend_notify_url]
+          missing = required_fields.select { |field| result_url[field.to_sym].nil? && result_url[field].nil? }
+
+          if missing.any?
+            raise ValidationError, "result_url must contain both frontend_redirect_url and backend_notify_url"
+          end
+        end
+
+        # Then validate frontend_redirect_url and backend_notify_url are present and not empty
+        if !options[:frontend_redirect_url].to_s.strip.empty? && !options[:backend_notify_url].to_s.strip.empty?
+          return
+        end
+
+        raise ValidationError, "result_url must contain both frontend_redirect_url and backend_notify_url"
+      end
+
       protected
 
       def payment_data
-        data = super
-        data[:result_url] = {
-          frontend_redirect_url: options[:frontend_redirect_url],
-          backend_notify_url: options[:backend_notify_url]
-        }
-        data[:prime] = options[:prime]
-        data
+        super.merge(
+          prime: options[:prime],
+          result_url: {
+            frontend_redirect_url: options[:frontend_redirect_url],
+            backend_notify_url: options[:backend_notify_url]
+          }
+        )
       end
     end
   end